refactor(api): split matrix.js — extract shared helpers + scope endpoints (Q1 part 2)
Stacked on the Q1 part-1 split. Behavior-preserving code moves:

- Shared helpers/constants (parseFilter, normaliseBlock, normaliseSortAttributes,
  resolveContextTypes, buildSubqueries, runCount, subjectScopeClauses, scopeCounts,
  getIdentityColumns/Values, ROW_TYPES/SAFE_IDENT_RE/FILTERABLE_TYPES) ->
  routes/matrix/shared.js. matrix.js imports the subset it still uses; call sites
  unchanged. normaliseSortAttributes re-exported from matrix.js for the test.
- Scope endpoints (/matrix/scope-stats, /scope-timeseries, /scope-breakdown) ->
  routes/matrix/scope.js, mounted via router.use() — public paths unchanged.
- Dropped 6 now-unused imports from matrix.js (randomUUID, buildEntitySubquery,
  collectContextIds, the 3 scopeHistory builders).

matrix.js: 1,417 -> 836 lines (1,679 before part 1). Adds shared.test.js (pure
helper unit tests). No query logic, routes, shapes, or ordering changed.
