+++ app/api/src/routes/matrix.js
@@ -14,7 +14,6 @@
 //   DELETE /api/matrix/saved-filters/:id   — remove
 
 import { Router } from 'express';
-import { randomUUID } from 'crypto';
 import * as db from '../db/connection.js';
 import { timedRequest } from '../perf/sqlTimer.js';
 import {
@@ -25,16 +24,7 @@ import {
   mergeGroupTotals,
   resourceMeta,
 } from '../db/matrixHelpers.js';
-import {
-  buildEntitySubquery,
-  collectContextIds,
-  UUID_RE,
-} from '../matrix/filterSql.js';
-import {
-  generateSampleDates,
-  buildScopeAsofSql,
-  historyStartSql,
-} from '../matrix/scopeHistory.js';
+import { UUID_RE } from '../matrix/filterSql.js';
 import {
   getPrincipalColumns,
   getResourceColumns,
@@ -52,267 +42,20 @@ import {
 import { buildAttrCutCellsSql, buildAttrCutNodesSql, tupleToNode } from '../matrix/attributeCut.js';
 import { buildRollupSql, buildRollupRolesSql, buildRolesAsRowsSql, buildGroupTotalsSql, buildRolesDrillSql } from '../matrix/rollupBuilders.js';
 import savedFiltersRouter from './matrix/savedFilters.js';
+import scopeRouter from './matrix/scope.js';
+import { getIdentityColumns, getIdentityColumnValues, parseFilter, buildSubqueries, runCount, subjectScopeClauses, scopeCounts } from './matrix/shared.js';
+// normaliseSortAttributes also lives in shared.js now; re-export so existing
+// consumers (matrix.rollup.test.js) keep importing it from here.
+export { normaliseSortAttributes } from './matrix/shared.js';
 
 const router = Router();
 // Saved-filter CRUD + the org-wide default filter live in their own module
 // (routes/matrix/savedFilters.js) — part of the matrix.js split (Q1).
 router.use(savedFiltersRouter);
+router.use(scopeRouter);
 const useSql = process.env.USE_SQL === 'true';
 
-const ROW_TYPES = new Set(['principal', 'identity']);
-const SAFE_IDENT_RE = /^[a-zA-Z0-9_]+$/;
-const FILTERABLE_TYPES = new Set([
-  'text', 'character varying', 'character', 'boolean',
-  'integer', 'bigint', 'smallint',
-]);
-
-// ─── Identity column discovery (small enough to live here) ──────────
-
-let identityColumnsCache = null;
-let identityColumnsCacheTime = 0;
-let identityValuesCache = null;
-let identityValuesCacheTime = 0;
-const IDENTITY_CACHE_TTL = 5 * 60 * 1000;
-
-async function getIdentityColumns() {
-  const now = Date.now();
-  if (identityColumnsCache && (now - identityColumnsCacheTime) < IDENTITY_CACHE_TTL) {
-    return identityColumnsCache;
-  }
-  const r = await db.query(
-    `SELECT column_name, data_type
-       FROM information_schema.columns
-      WHERE table_schema = 'public' AND table_name = 'Identities'
-        AND column_name NOT IN ('id', 'extendedAttributes')
-      ORDER BY ordinal_position`
-  );
-  identityColumnsCache = r.rows.map(row => ({
-    name: row.column_name,
-    rawName: row.column_name,
-    type: row.data_type,
-  }));
-  identityColumnsCacheTime = now;
-  return identityColumnsCache;
-}
-
-async function getIdentityColumnValues() {
-  const now = Date.now();
-  if (identityValuesCache && (now - identityValuesCacheTime) < IDENTITY_CACHE_TTL) {
-    return identityValuesCache;
-  }
-  const grouped = {};
-
-  // Distinct values for real, filterable columns.
-  const cols = await getIdentityColumns();
-  const filterable = cols.filter(c => FILTERABLE_TYPES.has(c.type) && SAFE_IDENT_RE.test(c.rawName));
-  if (filterable.length > 0) {
-    const parts = filterable.map(c =>
-      `SELECT '${c.name}' AS col, val FROM (
-         SELECT DISTINCT "${c.rawName}"::text AS val FROM "Identities"
-          WHERE "${c.rawName}" IS NOT NULL AND "${c.rawName}"::text <> ''
-          LIMIT 500
-       ) t`
-    );
-    const r = await db.query(parts.join('\nUNION ALL\n') + '\nORDER BY col, val');
-    for (const row of r.rows) {
-      if (!grouped[row.col]) grouped[row.col] = [];
-      grouped[row.col].push(row.val);
-    }
-  }
-
-  // Extension-attribute keys + distinct values, surfaced as ext.<key> so they
-  // can be picked and filtered just like Principal/Resource ext attributes.
-  try {
-    const ext = await db.query(`
-      SELECT col, val FROM (
-        SELECT DISTINCT 'ext.' || e.key AS col, e.value AS val
-          FROM "Identities" i, LATERAL jsonb_each_text(i."extendedAttributes") e
-         WHERE i."extendedAttributes" IS NOT NULL
-           AND e.value IS NOT NULL AND e.value <> ''
-      ) t
-      ORDER BY col, val
-      LIMIT 5000
-    `);
-    for (const row of ext.rows) {
-      if (!grouped[row.col]) grouped[row.col] = [];
-      if (grouped[row.col].length < 500) grouped[row.col].push(row.val);
-    }
-  } catch { /* extendedAttributes column may be absent on older schemas */ }
-
-  identityValuesCache = grouped;
-  identityValuesCacheTime = now;
-  return identityValuesCache;
-}
-
-// ─── Helpers ────────────────────────────────────────────────────────
-
-function parseFilter(body) {
-  const f = body && body.filter;
-  if (!f || typeof f !== 'object') return null;
-  const rowType = ROW_TYPES.has(f.rowType) ? f.rowType : 'principal';
-  return {
-    rowType,
-    subject:  normaliseBlock(f.subject),
-    resource: normaliseBlock(f.resource),
-    // Roll-up: aggregate the subject axis by this attribute (real column or
-    // ext.<key>). null = off. Validated against real columns in the handler.
-    rollup: typeof f.rollup === 'string' && f.rollup ? f.rollup : null,
-    // What the roll-up returns: resources + role columns (default), resources
-    // only, or business roles as rows.
-    rollupContent: ['resources-and-roles', 'resources-only', 'roles-only'].includes(f.rollupContent)
-      ? f.rollupContent : 'resources-and-roles',
-    // How each roll-up cell is displayed: an absolute count (default) or the
-    // percentage of in-scope subjects in that group who hold it. Presentational
-    // only — the backend always returns groupTotals so the frontend can switch.
-    rollupMetric: f.rollupMetric === 'percent' ? 'percent' : 'count',
-    // Internal: a roles-only drill request returns the per-subject breakdown
-    // for the group already scoped via a subject attribute condition.
-    drill: f.drill === true,
-    // EXPERIMENTAL — aggregate the subject axis by a Context tree (e.g. the
-    // Manager Hierarchy) instead of an attribute. rollupKind 'context' switches
-    // the roll-up branch; rollupContextId is the starting (root) node. The view
-    // zooms one level at a time: rollupPath is the drill path from the root to
-    // the current focus node, and the columns are the focus node's children.
-    rollupKind: f.rollupKind === 'context' ? 'context' : 'attribute',
-    rollupContextId: typeof f.rollupContextId === 'string' && f.rollupContextId ? f.rollupContextId : null,
-    rollupPath: Array.isArray(f.rollupPath) ? f.rollupPath.filter(x => typeof x === 'string').slice(0, 50) : [],
-    // Layered hierarchy view: the set of org nodes the user has expanded in
-    // place. The visible columns are the resulting tree cut; expanding a node
-    // adds the next level as a new header row (see buildContextCutSql). Reused by
-    // the layered attribute fold, where the entries are attribute-tuple keys.
-    rollupExpanded: Array.isArray(f.rollupExpanded) ? f.rollupExpanded.filter(x => typeof x === 'string').slice(0, 200) : [],
-    // Serve the attribute fold as a server-aggregated layered view (counts +
-    // expand-in-place) rather than a flat per-subject grid — set by the wizard
-    // for matrices too large to ship every row. Uses sortAttributes as the tree.
-    foldAttributes: f.foldAttributes === true,
-    // Layered attribute fold: the set of attribute-tuple keys the user has
-    // FOLDED (collapse model — default none = full depth, all attribute rows
-    // shown). Inverse of rollupExpanded, which the hierarchy view uses.
-    rollupCollapsed: Array.isArray(f.rollupCollapsed) ? f.rollupCollapsed.filter(x => typeof x === 'string').slice(0, 500) : [],
-    // Subject-axis sort order — client-side only, but normalised here so the
-    // shape is consistent across endpoints. Max 3 attributes.
-    sortAttributes: normaliseSortAttributes(f.sortAttributes),
-    // Sort the subject axis by a Context tree (Manager Hierarchy). Served as a
-    // context roll-up (aggregated per org node) so we never ship every
-    // per-subject row — see the translation in the /matrix/data handler.
-    sortHierarchy: f.sortHierarchy && typeof f.sortHierarchy === 'object'
-      && typeof f.sortHierarchy.contextId === 'string' && f.sortHierarchy.contextId
-      ? { contextId: f.sortHierarchy.contextId } : null,
-  };
-}
-
-export function normaliseSortAttributes(arr) {
-  const DEFAULT = [{ attribute: 'department', dir: 'asc' }];
-  if (!Array.isArray(arr)) return DEFAULT;
-  const out = [];
-  for (const a of arr) {
-    if (!a || typeof a.attribute !== 'string' || !a.attribute) continue;
-    out.push({ attribute: a.attribute, dir: a.dir === 'desc' ? 'desc' : 'asc' });
-    if (out.length === 6) break;
-  }
-  return out.length ? out : DEFAULT;
-}
-
-function normaliseBlock(b) {
-  if (!b || typeof b !== 'object') return { include: [], exclude: [] };
-  return {
-    include: Array.isArray(b.include) ? b.include : [],
-    exclude: Array.isArray(b.exclude) ? b.exclude : [],
-  };
-}
-
-
-async function resolveContextTypes(filter) {
-  const ids = collectContextIds(filter);
-  if (ids.length === 0) return new Map();
-  const r = await db.query(
-    `SELECT id, "targetType" FROM "Contexts" WHERE id = ANY($1::uuid[])`,
-    [ids],
-  );
-  return new Map(r.rows.map(row => [row.id, row.targetType]));
-}
-
-async function buildSubqueries(filter) {
-  const [principalCols, resourceCols, identityCols, contextTypes] = await Promise.all([
-    getPrincipalColumns(),
-    getResourceColumns(),
-    getIdentityColumns(),
-    resolveContextTypes(filter),
-  ]);
-  const principalColSet = new Set(principalCols.map(c => c.name));
-  const resourceColSet  = new Set(resourceCols.map(c => c.name));
-  const identityColSet  = new Set(identityCols.map(c => c.name));
-
-  const subjectEntity = filter.rowType === 'identity' ? 'Identity' : 'Principal';
-  const subjectValidCols = filter.rowType === 'identity' ? identityColSet : principalColSet;
-
-  const subject = buildEntitySubquery({
-    entity: subjectEntity,
-    include: filter.subject.include,
-    exclude: filter.subject.exclude,
-    validColumns: subjectValidCols,
-    contextTypes,
-    bindingPrefix: 'sf',
-  });
-  const resource = buildEntitySubquery({
-    entity: 'Resource',
-    include: filter.resource.include,
-    exclude: filter.resource.exclude,
-    validColumns: resourceColSet,
-    contextTypes,
-    bindingPrefix: 'rf',
-  });
-
-  return {
-    subjectSql:    subject.sql,
-    resourceSql:   resource.sql,
-    bindings:      { ...subject.bindings, ...resource.bindings },
-    warnings:      [...subject.warnings, ...resource.warnings],
-    principalCols,
-    resourceCols,
-    identityCols,
-  };
-}
-
-// Run a single-cell COUNT query with the given bindings; returns the integer.
-async function runCount(p, label, res, sql, bindings) {
-  const r = timedRequest(p, label, res);
-  for (const [k, v] of Object.entries(bindings)) r.input(k, v);
-  const result = await r.query(sql);
-  return result.recordset[0]?.c ?? 0;
-}
-
-function subjectScopeClauses(rowType, subjectSql) {
-  const subjectTable = rowType === 'identity' ? 'Identities' : 'Principals';
-  // For principals, exclude group-shaped accounts so counts match what the
-  // matrix actually renders.
-  const baseWhere = rowType === 'principal'
-    ? `("principalType" IS NULL OR "principalType" != '#microsoft.graph.group')`
-    : null;
-  const idClause = subjectSql ? `id IN ${subjectSql}` : null;
-  const clauses = [baseWhere, idClause].filter(Boolean);
-  return {
-    subjectTable,
-    where: clauses.length ? ` WHERE ${clauses.join(' AND ')}` : '',
-    baseWhere: baseWhere ? ` WHERE ${baseWhere}` : '',
-  };
-}
-
-// Subject/resource scope counts shared by /matrix/data (flat + roll-up paths).
-async function scopeCounts(p, res, rowType, built) {
-  const subj = subjectScopeClauses(rowType, built.subjectSql);
-  const [subjectCount, subjectTotal, resourceCount, resourceTotal] = await Promise.all([
-    runCount(p, 'matrix-data-subject-count', res,
-      `SELECT COUNT(*)::int AS c FROM "${subj.subjectTable}"${subj.where}`, built.bindings),
-    runCount(p, 'matrix-data-subject-total', res,
-      `SELECT COUNT(*)::int AS c FROM "${subj.subjectTable}"${subj.baseWhere}`, {}),
-    runCount(p, 'matrix-data-resource-count', res,
-      `SELECT COUNT(*)::int AS c FROM "Resources"${built.resourceSql ? ` WHERE id IN ${built.resourceSql}` : ''}`, built.bindings),
-    runCount(p, 'matrix-data-resource-total', res,
-      `SELECT COUNT(*)::int AS c FROM "Resources"`, {}),
-  ]);
-  return { subjectCount, subjectTotal, resourceCount, resourceTotal };
-}
+// Shared helpers + constants moved to ./matrix/shared.js (Q1 split); imported below.
 
 // Roll-up SQL builders moved to ../matrix/rollupBuilders.js (Q1 split); they are
 // imported above and used by the handlers below. Their unit tests live in
@@ -371,247 +114,6 @@ router.post('/matrix/preview', async (req, res) => {
   } catch (err) {
     console.error('matrix/preview failed:', err.message);
     return res.status(500).json({ error: 'Preview failed' });
-  }
-});
-
-// ─── POST /api/matrix/scope-stats ───────────────────────────────────
-// Richer counts for the current selection than /preview: subject + resource
-// totals PLUS the governed-vs-non-governed assignment split. Powers the Scope
-// Statistics panel. "Governed" mirrors the matrix's own managed/SOLL semantics: a
-// (principal, resource) pair is governed when the membership is covered by a business
-// role the user holds — it appears in vw_UserPermissionAssignmentViaBusinessRole
-// (a BusinessRole that Contains the group AND a Governed assignment of that role to the
-// user). This is exactly what the matrix colours as SOLL.
-router.post('/matrix/scope-stats', async (req, res) => {
-  if (!useSql) {
-    return res.json({
-      subjectCount: 0, resourceCount: 0,
-      assignmentCount: 0, governedAssignmentCount: 0, ungovernedAssignmentCount: 0,
-      governedPct: 0, rowType: 'principal',
-    });
-  }
-  const filter = parseFilter(req.body);
-  if (!filter) return res.status(400).json({ error: 'Invalid filter body' });
-
-  try {
-    const built = await buildSubqueries(filter);
-    const p = await db.getPool();
-    const subj = subjectScopeClauses(filter.rowType, built.subjectSql);
-
-    const { subjectIdExpr, assignmentJoin, assignmentWhere } = buildAssignmentExprs(filter.rowType, built);
-
-    // One pair-level aggregation: distinct (subject, resource) pairs, each
-    // flagged governed if ANY of its assignment rows is access-package managed.
-    const pairReq = timedRequest(p, 'matrix-scope-pairs', res);
-    for (const [k, v] of Object.entries(built.bindings)) pairReq.input(k, v);
-    const pairSql = `
-      SELECT
-        COUNT(*)::int AS total,
-        COUNT(*) FILTER (WHERE managed)::int AS governed
-      FROM (
-        SELECT ${subjectIdExpr} AS sid, p."resourceId" AS rid,
-               bool_or(br."userId" IS NOT NULL) AS managed
-          FROM "vw_ResourceUserPermissionAssignments" p
-          ${assignmentJoin}
-          LEFT JOIN "vw_UserPermissionAssignmentViaBusinessRole" br
-            ON br."userId" = p."principalId" AND br."resourceId" = p."resourceId"
-         WHERE ${assignmentWhere.join(' AND ')}
-         GROUP BY ${subjectIdExpr}, p."resourceId"
-      ) pairs`;
-
-    const [subjectCount, resourceCount, pairRow] = await Promise.all([
-      runCount(p, 'matrix-scope-subject', res,
-        `SELECT COUNT(*)::int AS c FROM "${subj.subjectTable}"${subj.where}`,
-        built.bindings),
-      runCount(p, 'matrix-scope-resource', res,
-        `SELECT COUNT(*)::int AS c FROM "Resources"${built.resourceSql ? ` WHERE id IN ${built.resourceSql}` : ''}`,
-        built.bindings),
-      pairReq.query(pairSql).then(r => r.recordset[0] || { total: 0, governed: 0 }),
-    ]);
-
-    const assignmentCount = pairRow.total || 0;
-    const governedAssignmentCount = pairRow.governed || 0;
-    const ungovernedAssignmentCount = assignmentCount - governedAssignmentCount;
-    const governedPct = assignmentCount > 0
-      ? Math.round((governedAssignmentCount / assignmentCount) * 1000) / 10
-      : 0;
-
-    return res.json({
-      rowType: filter.rowType,
-      subjectCount,
-      resourceCount,
-      assignmentCount,
-      governedAssignmentCount,
-      ungovernedAssignmentCount,
-      governedPct,
-      warnings: built.warnings,
-    });
-  } catch (err) {
-    console.error('matrix/scope-stats failed:', err.message);
-    return res.status(500).json({ error: 'Scope statistics failed' });
-  }
-});
-
-// ─── POST /api/matrix/scope-timeseries ──────────────────────────────
-// Historic timeline for the current selection, reconstructed from the audit
-// log (_history) — no dedicated snapshot tables. Body: { filter }. Query:
-// days (range, default 180), points (samples, default 13). Returns one point
-// per sample date with principals / resources / assignments / governed counts
-// and the governed %, plus `historyStart` (earliest reliable instant) and
-// `scopeMode` ('attribute' = fully reconstructed; 'context-current' = scope
-// membership taken from today because ContextMembers is not audited).
-router.post('/matrix/scope-timeseries', async (req, res) => {
-  if (!useSql) return res.json({ historyStart: null, scopeMode: 'attribute', points: [] });
-
-  const filter = parseFilter(req.body);
-  if (!filter) return res.status(400).json({ error: 'Invalid filter body' });
-
-  try {
-    const [principalCols, resourceCols, contextTypes] = await Promise.all([
-      getPrincipalColumns(),
-      getResourceColumns(),
-      resolveContextTypes(filter),
-    ]);
-    const principalColSet = new Set(principalCols.map(c => c.name));
-    const resourceColSet  = new Set(resourceCols.map(c => c.name));
-
-    const { sql, bindings, warnings, scopeMode } = buildScopeAsofSql({
-      filter, principalColSet, resourceColSet, contextTypes,
-    });
-
-    const p = await db.getPool();
-
-    // The audit log is pruned beyond HISTORY_RETENTION_DAYS (default 180), so
-    // the timeline can't reach further back than that. Clamp the requested
-    // range and tell the client the bound so the UI can explain it.
-    const retRow = await db.queryOne(
-      `SELECT "configValue" FROM "WorkerConfig" WHERE "configKey" = 'HISTORY_RETENTION_DAYS'`,
-    );
-    const retentionDays = retRow ? parseInt(retRow.configValue, 10) : 180;
-    const reqDays = parseInt(req.query.days, 10) || 180;
-    const days = (retentionDays > 0) ? Math.min(reqDays, retentionDays) : reqDays;
-
-    // Earliest reliable instant — don't fabricate points before auditing began.
-    const startRow = await db.queryOne(historyStartSql());
-    const historyStart = startRow?.start ? new Date(startRow.start) : null;
-
-    const dates = generateSampleDates({ days, points: req.query.points });
-
-    const points = await Promise.all(dates.map(async (date) => {
-      const asof = `${date}T23:59:59.999Z`;
-      // Skip points before auditing began (reconstruction unreliable there).
-      if (historyStart && new Date(asof) < historyStart) {
-        return { date, principals: 0, resources: 0, assignments: 0, governed: 0, governedPct: 0, beforeHistory: true };
-      }
-      const r = timedRequest(p, 'matrix-scope-timeseries', res);
-      for (const [k, v] of Object.entries(bindings)) r.input(k, v);
-      r.input('asof', asof);
-      const row = (await r.query(sql)).recordset[0] || {};
-      const assignments = row.assignments || 0;
-      const governed = row.governed || 0;
-      return {
-        date,
-        principals: row.principals || 0,
-        resources: row.resources || 0,
-        assignments,
-        governed,
-        ungoverned: assignments - governed,
-        governedPct: assignments > 0 ? Math.round((governed / assignments) * 1000) / 10 : 0,
-      };
-    }));
-
-    return res.json({
-      historyStart: historyStart ? historyStart.toISOString() : null,
-      retentionDays,
-      scopeMode,
-      points,
-      warnings,
-    });
-  } catch (err) {
-    console.error('matrix/scope-timeseries failed:', err.message);
-    return res.status(500).json({ error: 'Scope timeseries failed' });
-  }
-});
-
-// ─── POST /api/matrix/scope-breakdown ───────────────────────────────
-// Department-by-department (or any principal attribute) breakdown of the
-// current selection: principals, assignments, and governed split per group.
-// Body: { filter }. Query: attribute (default 'department', or an ext.* key).
-// The UI drills into a group by re-querying scope-timeseries with that
-// attribute added to the subject filter.
-router.post('/matrix/scope-breakdown', async (req, res) => {
-  if (!useSql) return res.json({ attribute: 'department', groups: [] });
-
-  const filter = parseFilter(req.body);
-  if (!filter) return res.status(400).json({ error: 'Invalid filter body' });
-
-  // Resolve + validate the grouping attribute (real principal column or ext.*).
-  const rawAttr = typeof req.query.attribute === 'string' && req.query.attribute
-    ? req.query.attribute : 'department';
-  const resolved = resolveAttrExpr(rawAttr, 'u', await getPrincipalColumns());
-  if (resolved.error) return res.status(400).json({ error: resolved.error });
-  const attrExpr = resolved.attrExpr;
-
-  try {
-    const built = await buildSubqueries(filter);
-    const p = await db.getPool();
-    const grp = `COALESCE(NULLIF(${attrExpr}::text, ''), '(none)')`;
-    const notGroup = `(u."principalType" IS NULL OR u."principalType" <> '#microsoft.graph.group')`;
-    const subjIn = built.subjectSql ? ` AND u.id IN ${built.subjectSql}` : '';
-    const resIn  = built.resourceSql ? ` AND p."resourceId" IN ${built.resourceSql}` : '';
-
-    // Principals per group (includes principals with no assignments).
-    const principalsReq = timedRequest(p, 'matrix-breakdown-principals', res);
-    for (const [k, v] of Object.entries(built.bindings)) principalsReq.input(k, v);
-    const principalsRows = (await principalsReq.query(`
-      SELECT ${grp} AS grp, COUNT(*)::int AS principals
-        FROM "Principals" u
-       WHERE ${notGroup}${subjIn}
-       GROUP BY ${grp}
-    `)).recordset;
-
-    // Assignment pairs + governed split per group.
-    const pairsReq = timedRequest(p, 'matrix-breakdown-pairs', res);
-    for (const [k, v] of Object.entries(built.bindings)) pairsReq.input(k, v);
-    const pairsRows = (await pairsReq.query(`
-      SELECT grp,
-             COUNT(*)::int AS assignments,
-             COUNT(*) FILTER (WHERE managed)::int AS governed
-        FROM (
-          SELECT ${grp} AS grp, u.id AS pid, p."resourceId" AS rid,
-                 bool_or(br."userId" IS NOT NULL) AS managed
-            FROM "Principals" u
-            JOIN "vw_ResourceUserPermissionAssignments" p ON p."principalId" = u.id
-            LEFT JOIN "vw_UserPermissionAssignmentViaBusinessRole" br
-              ON br."userId" = u.id AND br."resourceId" = p."resourceId"
-           WHERE ${notGroup}${subjIn}${resIn}
-           GROUP BY ${grp}, u.id, p."resourceId"
-        ) t
-       GROUP BY grp
-    `)).recordset;
-
-    // Merge the two result sets by group key.
-    const byGroup = new Map();
-    for (const r of principalsRows) {
-      byGroup.set(r.grp, { group: r.grp, principals: r.principals, assignments: 0, governed: 0 });
-    }
-    for (const r of pairsRows) {
-      const g = byGroup.get(r.grp) || { group: r.grp, principals: 0, assignments: 0, governed: 0 };
-      g.assignments = r.assignments;
-      g.governed = r.governed;
-      byGroup.set(r.grp, g);
-    }
-
-    const groups = [...byGroup.values()].map(g => ({
-      ...g,
-      ungoverned: g.assignments - g.governed,
-      governedPct: g.assignments > 0 ? Math.round((g.governed / g.assignments) * 1000) / 10 : 0,
-    })).sort((a, b) => b.assignments - a.assignments);
-
-    return res.json({ attribute: rawAttr, groups, warnings: built.warnings });
-  } catch (err) {
-    console.error('matrix/scope-breakdown failed:', err.message);
-    return res.status(500).json({ error: 'Scope breakdown failed' });
   }
 });
